[Feat] Add generic nodeId to allow connecting to any node in a cluster
This commit adds a new nodeId value (-2) that has special behavior when
used in `clusters.json`.

Typically, we expect each cluster in `clusters.json` to have a
well-defined set of hosts and nodeIds. As a result, every change in
cluster topology/memebership requires updating `clusters.json`.

Instead, a cluster can contain a single node in `clusters.json` which
uses the new, generic nodeId contains a single node with a DNS endpoint
that resolves to the list of IPs for the cluster.